+++ source/include/Ostap/Differences.h
@@ -131,11 +131,223 @@ namespace Ostap
       // ======================================================================
       // Finite differences
       // ======================================================================
+      /** @class Difference
+       *  Finite difference
+       *  https://en.wikipedia.org/wiki/Finite_difference
+       *  @author Vanya BELYAEV Ivan.Belyaev@cern.ch
+       */
+      class FiniteDifference_
+      {
+	// ====================================================================
+      public : // N = 0 
+	// ====================================================================
+	/// forward difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==0),int>::type = 0 > 
+	static inline double forward
+	( const FUNCTION& fun ,
+	  const double    x   ,
+	  const double    h   = 1 ) { return fun ( x ) ; }
+	// ====================================================================
+	/// backward difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==0),int>::type = 0 > 
+	static inline double backward 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) { return fun ( x ) ; }
+	// ====================================================================
+	/// central difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==0),int>::type = 0 > 
+	static inline double central 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) { return fun ( x ) ; }
+	// ====================================================================
+      public : // N == 1 
+	// ====================================================================
+	/// forward difference 
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==1),int>::type = 0 > 
+	static inline double forward
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) { return !h ? 0.0 : fun ( x + 1.0L * h ) - 1.0L * fun ( x ) ; }
+	// ====================================================================
+	/// backward difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==1),int>::type = 0 > 
+	static inline double backward 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) { return !h ? 0.0 : fun ( x ) - 1.0L * fun ( x - 1.0L * h ) ; }
+	// ====================================================================
+	/// central difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==1),int>::type = 0 > 
+	static inline double central 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) { return !h ? 0.0 : fun ( x + 0.5L * h ) - 1.0L * fun ( x - 0.5L * h ) ; }
+	// ====================================================================
+      public : // N == 2 
+	// ====================================================================
+	/// forward difference 
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==2),int>::type = 0 > 
+	static inline double forward
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )
+	{ return !h ? 0.0 : fun ( x ) - 2.0L * fun ( x + 1.0L * h ) + fun ( x + 2.0L * h )  ; }
+	// ====================================================================
+	/// backward difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==2),int>::type = 0 > 
+	static inline double backward 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )
+	{ return !h ? 0.0 : fun ( x ) - 2.0L * fun ( x - 1.0L * h ) + fun ( x - 2.0L * h )  ; }
+	// ====================================================================
+	/// central difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==2),int>::type = 0 > 
+	static inline double central 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) 
+	{ return !h ? 0.0 : fun ( x + 1.0L * h ) - 2.0L * fun ( x ) + fun ( x - 1.0L * h )  ; }
+	// ====================================================================
+      public : // N == 3 
+	// ====================================================================
+	/// forward difference 
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==3),int>::type = 0 > 
+	static inline double forward
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )
+	{ return !h ? 0.0 : - fun ( x ) + 3.0L * fun ( x + 1.0L * h ) - 3.0L * fun (  x + 2.0L * h ) + fun ( x + 3.0L * h ) ; }
+	// ====================================================================
+	/// backward difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==3),int>::type = 0 > 
+	static inline double backward 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )
+	{ return !h ? 0.0 : + fun ( x ) - 3.0L * fun ( x - 1.0L * h ) + 3.0L * fun (  x - 2.0L * h ) - fun ( x - 3.0L * h ) ; }
+	// ====================================================================
+	/// central difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N==3),int>::type = 0 > 
+	static inline double central 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 ) 
+	{ return !h ? 0.0 : + fun ( x + 1.5L * h ) - 3.0L * fun ( x + 0.5L * h ) + 3.0L * fun ( x - 0.5L * h ) - fun ( x - 1.5L * h ) ; } 
+	// ====================================================================
+      public : // N > 3 
+	// ====================================================================
+	/// forward difference 
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N>3),int>::type = 0 > 
+	static inline double forward
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )	
+	{
+	  if ( !h ) { return 0 ; } 
+	  long double r = fun ( x ) * ( 0 == N % 2 ? 1 : -1 ) ;
+	  long double t = 1 ;
+	  for ( unsigned short j = 1 ; j <= N ; ++j )
+	  {
+	    t *= - ( N + 1 - j ) ;
+	    t /= j ;
+	    r += t * fun ( x + 1.0L * j * h ) ; 
+	  }
+	  return r ;
+	}
+	// ====================================================================
+	/// backward difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N>3),int>::type = 0 > 
+	static inline double backward 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )
+	{
+	  if ( !h ) { return 0 ; } 
+	  long double r = fun ( x ) ;
+	  long double t = 1 ;
+	  for ( unsigned short j = 1 ; j <= N ; ++j )
+	  {
+	    t *= - ( N + 1 - j ) ;
+	    t /= j ;
+	    r += t * fun ( x - 1.0L * j * h ) ; 
+	  }
+	  return r ;
+	}	
+	// ====================================================================
+	/// central difference
+	template <class FUNCTION,
+		  const unsigned short N, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 ,	  
+		  typename std::enable_if<(N>3),int>::type = 0 > 
+	static inline double central 
+	( const FUNCTION& fun     ,
+	  const double    x       ,
+	  const double    h   = 1 )
+	{
+	  if ( !h ) { return 0 ; } 
+	  long double r = fun ( x ) ;
+	  long double t = 1 ;
+	  for ( unsigned short j = 1 ; j <= N ; ++j )
+	  {
+	    t *= - ( N + 1 - j ) ;
+	    t /= j ;
+	    r += t * fun ( x + ( 0.5L * N - j ) * h ) ; 
+	  }
+	  return r ;
+	}	  
+	// ====================================================================	
+      };
+      // ======================================================================
       /** @class Forward
        *  simple evaluator of the Nth forward difference
        *  @see https://en.wikipedia.org/wiki/Finite_difference
        */
-      template <unsigned int N> 
+      template <unsigned short N> 
       class Forward
       {
       public:
@@ -146,50 +358,21 @@ namespace Ostap
          *  @param h   step 
          *  @return Nth difference 
          */
-        template <class FUNCTION>
-        inline 
-        static double evaluate 
-        ( FUNCTION     fun , 
-          const double x   , 
-          const double h   ) ;
+	template <class FUNCTION,
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0>
+        static inline double evaluate 
+        ( const FUNCTION& fun     , 
+          const double    x       , 
+          const double    h   = 1 )
+	{ return FiniteDifference_::forward<N> ( fun , x , h ) ; }
         // ====================================================================
       } ;
-      // ======================================================================
-      /// specialization for the 0th difference (stop recursion)
-      template <> 
-      class Forward<0>
-      {
-      public:
-        // ====================================================================
-        template <class FUNCTION>
-        inline 
-        static double evaluate 
-        ( FUNCTION        fun  , 
-          const double    x    , 
-          const double /* h */ ) { return fun ( x ) ; }
-        // ====================================================================
-      };
-      // ====================================================================== 
-      /// recursive method to get the Nth forward difference 
-      template <unsigned int N> 
-      template <class FUNCTION> 
-      inline 
-      double Forward<N>::evaluate 
-      ( FUNCTION fun      , 
-        const    double x , 
-        const    double h ) 
-      {
-        const auto cfunr = std::cref ( fun ) ;
-        return 
-          Forward<N-1>::evaluate ( cfunr , x + h , h ) - 
-          Forward<N-1>::evaluate ( cfunr , x     , h ) ;
-      }
       // ======================================================================
       /** @class Backward
        *  simple evaluator of the Nth backward difference
        *  @see https://en.wikipedia.org/wiki/Finite_difference
        */
-      template <unsigned int N> 
+      template <unsigned short N> 
       class Backward
       {
       public:
@@ -200,162 +383,248 @@ namespace Ostap
          *  @param h   step 
          *  @return Nth difference 
          */
-        template <class FUNCTION>
-        inline 
-        static double evaluate
-        ( FUNCTION     fun , 
-          const double x   , 
-          const double h   ) ;
+	template <class FUNCTION,
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0>
+        static inline double evaluate 
+        ( const FUNCTION& fun     , 
+          const double    x       , 
+          const double    h   = 1 )
+	{ return FiniteDifference_::backward<N> ( fun , x , h ) ; }
         // ====================================================================
       } ;
-      // ======================================================================
-      /// specialization for the 0th difference (stop recursion)
-      template <> 
-      class Backward<0>
-      {
-      public:
-        // ====================================================================
-        template <class FUNCTION>
-        inline 
-        static double evaluate 
-        ( FUNCTION        fun  , 
-          const double    x    , 
-          const double /* h */ ) { return fun ( x ) ; }
-        // ====================================================================
-      };
-      // ====================================================================== 
-      /// recursive method to get the Nth forward difference 
-      template <unsigned int N> 
-      template <class FUNCTION> 
-      inline 
-      double Backward<N>::evaluate
-      ( FUNCTION fun      , 
-        const    double x , 
-        const    double h ) 
-      {
-        const auto cfunr = std::cref ( fun ) ;
-        return 
-          Forward<N-1>::evaluate ( cfunr , x     , h ) - 
-          Forward<N-1>::evaluate ( cfunr , x - h , h ) ;
-      }
       // ======================================================================
       /** @class Central
        *  simple evaluator of the Nth central difference
        *  @see https://en.wikipedia.org/wiki/Finite_difference
        */
-      template <unsigned int N> 
+      template <unsigned short N> 
       class Central 
       {
       public:
-        // ====================================================================
+        // ==================================================================== 
         /** get Nth difference
          *  @param fun the fnuction
          *  @param x   x-value 
          *  @param h   step 
          *  @return Nth difference 
          */
-        template <class FUNCTION>
-        inline 
-        static double evaluate
-        ( FUNCTION     fun , 
-          const double x   , 
-          const double h   ) ;
+	template <class FUNCTION,
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0>
+        static inline double evaluate 
+        ( const FUNCTION& fun     , 
+          const double    x       , 
+          const double    h   = 1 )
+	{ return FiniteDifference_::central<N> ( fun , x , h ) ; }
         // ====================================================================
       } ;
-      // ======================================================================
-      /// specialization for the 0th difference (stop recursion)
-      template <> 
-      class Central <0>
-      {
-      public:
-        // ====================================================================
-        template <class FUNCTION>
-        inline 
-        static double evaluate 
-        ( FUNCTION        fun  , 
-          const double    x    , 
-          const double /* h */ ) { return fun ( x ) ; }
-        // ====================================================================
-      };
-      // ====================================================================== 
-      /// recursive method to get the Nth forward difference 
-      template <unsigned int N> 
-      template <class FUNCTION> 
-      inline 
-      double Central<N>::evaluate 
-      ( FUNCTION fun      , 
-        const    double x , 
-        const    double h ) 
-      {
-        const auto cfunr = std::cref ( fun ) ;
-        return 
-          Central<N-1>::evaluate ( cfunr , x     , h ) - 
-          Central<N-1>::evaluate ( cfunr , x - h , h ) ;
-      }
       // =======================================================================
-      /** recursively evaluate N-th forward difference of function <code>fun</code>
+      /** Evaluate N-th forward difference of function <code>fun</code>
        *  @param fun the function 
        *  @param N   the order of the difference 
        *  @param x   the point 
        *  @param h   the step 
        *  @return N-th forward dirrerence 
        */
-      template <class FUNCTION>
-      inline double forward 
-      ( FUNCTION             fun , 
-        const unsigned short N   , 
-        const double         x   , 
-        const double         h   ) 
+      template <class FUNCTION,
+		typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0>
+      inline double forward_ 
+      ( const FUNCTION&      fun     , 
+        const unsigned short N       , 
+        const double         x       , 
+        const double         h   = 1 ) 
       {
-        const auto cfunr = std::cref ( fun );
-        return 
-          0 == N ? fun ( x ) :
-          forward ( cfunr , N - 1 , x + h , h ) - 
-          forward ( cfunr , N - 1 , x     , h ) ;  
+	//
+	if      ( 0 == N ) { return fun ( x ) ; }
+	else if ( !h     ) { return 0         ; }
+	else if ( 1 == N ) { return fun ( x + h ) - 1.0L * fun ( x ) ; }
+	else if ( 2 == N ) { return fun ( x )     - 2.0L * fun ( x + h ) + fun ( x + 2 * h ) ; }
+	//
+	long double r = fun ( x ) * ( 0 == N % 2 ? 1 : -1 ) ;
+	long double t = 1 ;
+	for ( unsigned short j = 1 ; j <= N ; ++j )
+	{
+	  t *= - ( N + 1 - j ) ;
+	  t /= j ;
+	  r += t * fun ( x + j * h ) ; 
+	}
+	return r ;	
       }
       // ======================================================================
-      /** recursively evaluate N-th backward difference of function <code>fun</code>
+      /** Evaluate N-th backward difference of function <code>fun</code>
        *  @param fun the function 
        *  @param N   the order of the difference 
        *  @param x   the point 
        *  @param h   the step 
        *  @return N-th backward dirrerence 
        */
-      template <class FUNCTION>
-      inline double backward 
-      ( FUNCTION             fun , 
-        const unsigned short N   , 
-        const double         x   , 
-        const double         h   ) 
+      template <class FUNCTION,
+		typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0>
+      inline double backward_ 
+      ( const FUNCTION&      fun     , 
+        const unsigned short N       , 
+        const double         x       , 
+        const double         h   = 1 ) 
       {
-        const auto cfunr = std::cref ( fun );
-        return 
-          0 == N ? fun ( x ) :
-          backward ( cfunr , N - 1 , x     , h ) - 
-          backward ( cfunr , N - 1 , x - h , h ) ;  
+	//
+	if      ( 0 == N ) { return fun ( x ) ; }
+	else if ( !h     ) { return 0         ; }
+	else if ( 1 == N ) { return fun ( x ) - 1.0L * fun ( x - h ) ; }
+	else if ( 2 == N ) { return fun ( x ) - 2.0L * fun ( x - h ) + fun ( x - 2 * h ) ; }
+	//
+	long double r = fun ( x ) ;
+	long double t = 1 ;
+	for ( unsigned short j = 1 ; j <= N ; ++j )
+	{
+	  t *= - ( N + 1 - j ) ;
+	  t /= j ;
+	  r += t * fun ( x - j * h ) ; 
+	}
+	return r ;	
       }
       // ======================================================================
-      /** recursivelyu evaluate N-th central difference of function <code>fun</code>
+      /** Evaluate N-th central difference of function <code>fun</code>
        *  @param fun the function 
        *  @param N   the order of the difference 
        *  @param x   the point 
        *  @param h   the step 
        *  @return N-th central dirrerence 
        */
-      template <class FUNCTION>
-      inline double central
-      ( FUNCTION             fun , 
-        const unsigned short N   , 
-        const double         x   , 
-        const double         h   ) 
+      template <class FUNCTION,
+		typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0>
+      inline double central_
+      ( const FUNCTION&      fun     , 
+        const unsigned short N       , 
+        const double         x       , 
+        const double         h   = 1 ) 
       {
-        const auto cfunr = std::cref ( fun );
-        return 
-          0 == N ? fun ( x ) :
-          central ( cfunr , N - 1 , x + 0.5 * h , h ) - 
-          central ( cfunr , N - 1 , x - 0.5 * h , h ) ;  
+	if      ( 0 == N ) { return fun ( x ) ; }
+	else if ( !h     ) { return 0         ; }
+	else if ( 1 == N ) { return fun ( x + 0.5L * h ) - 1.0L * fun ( x - 0.5L * h ) ; }
+	else if ( 2 == N ) { return fun ( x + 1.0L * h ) - 2.0L * fun ( x ) + fun ( x - 1.0L * h ) ; }
+	//
+	long double r = fun ( x ) ;
+	long double t = 1 ;
+	for ( unsigned short j = 1 ; j <= N ; ++j )
+	  {
+	  t *= - ( N + 1 - j ) ;
+	  t /= j ;
+	  r += t * fun ( x - j * h ) ; 
+	}
+	return r ;
       }
       // ======================================================================
+      /** Evaluate N-th forward difference of function <code>fun</code>
+       *  @param fun the function 
+       *  @param N   the order of the difference 
+       *  @param x   the point 
+       *  @param h   the step 
+       *  @return N-th forward dirrerence 
+       */
+      double forward 
+      ( std::function<double(double)> fun     , 
+        const unsigned short          N       , 
+        const double                  x       , 
+        const double                  h   = 1 ) ;
+      // ======================================================================
+      /** Evaluate N-th backward difference of function <code>fun</code>
+       *  @param fun the function 
+       *  @param N   the order of the difference 
+       *  @param x   the point 
+       *  @param h   the step 
+       *  @return N-th backward dirrerence 
+       */
+      double backward 
+      ( std::function<double(double)> fun , 
+        const unsigned short          N     , 
+        const double                  x     , 
+        const double                  h = 1 ) ;
+      // ======================================================================
+      /** Evaluate N-th central difference of function <code>fun</code>
+       *  @param fun the function 
+       *  @param N   the order of the difference 
+       *  @param x   the point 
+       *  @param h   the step 
+       *  @return N-th central dirrerence 
+       */
+      double central
+      ( std::function<double(double)> fun   , 
+        const unsigned short          N     , 
+        const double                  x     , 
+        const double                  h = 1 ) ;
+      // ======================================================================
+      /** @class FiniteDifference 
+       */
+      class FiniteDifference
+      {
+      public :
+	// ===================================================================
+	/** construcructor for the order 
+	 *  @param N  the order of the finite differenece 
+	 */
+	FiniteDifference
+	( const unsigned short N = 1 ) ;
+	// =================================================================
+      public :
+	// =================================================================
+	/// get forward difference 
+	template <class FUNCTION, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 >       	
+	inline double forward 
+	( const FUNCTION& f     ,
+	  const double    x     ,
+	  const double    h = 1 ) const
+	{ return Ostap::Math::Differences::forward_ ( f , m_N , x , h ) ;  }
+	// =================================================================
+	/// get backward difference 
+	template <class FUNCTION, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 >       	
+	inline double backward 
+	( const FUNCTION& f     ,
+	  const double    x     ,
+	  const double    h = 1 ) const
+	{ return Ostap::Math::Differences::backward_ ( f , m_N , x , h ) ;  }
+	// =================================================================
+	/// get central difference 
+	template <class FUNCTION, 
+		  typename std::enable_if<std::is_invocable<FUNCTION,double>::value,int>::type = 0 >       	
+	inline double central 
+	( const FUNCTION& f     ,
+	  const double    x     ,
+	  const double    h = 1 ) const
+	{ return Ostap::Math::Differences::central_ ( f , m_N , x , h ) ;  }
+	// =================================================================	
+      public :
+	// =================================================================
+	/// get forward difference 
+	double forward
+	( std::function<double(double)> f     , 
+	  const double                  x     ,
+	  const double                  h = 1 ) const ;
+	// =================================================================
+	/// get backward difference 
+	double backward 
+	( std::function<double(double)> f     , 
+	  const double                  x     ,
+	  const double                  h = 1 ) const ;
+	// =================================================================
+	/// get central difference 
+	double central 
+	( std::function<double(double)> f     , 
+	  const double                  x     ,
+	  const double                  h = 1 ) const ;
+	// =================================================================	
+      public :
+	// =================================================================
+	/// get the order 
+	inline unsigned short N () const { return m_N ; }
+	// =================================================================
+      private :
+	// =================================================================
+	/// the order :
+	unsigned short m_N { 1 } ; // the order 
+	// =================================================================	
+      } ;		
+      // ======================================================================      
     } //                          The end of namespace Ostap::Math::Differences  
     // ========================================================================
   } //                                         The end of namespace Ostap::Math